remade fetch to return a u32 instead of a i64

+++ Emulator/include/pipeline/fetch.h
@@ -4,6 +4,6 @@
 #include "my_stdtypes.h"
 #include "emulator/emulator.h"
 
-i64 fetch(Emulator *emu);
+u32 fetch(Emulator *emu);
 
 #endif
